requirements.txt
Django==2.0.2
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2017.3
djangorestframework
